requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.3.0
werkzeug==3.0.3
zipp==3.19.2
